requirements.txt
openai>=1.30.0
python-docx>=1.1.0
PyPDF2>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
streamlit>=1.34.0
